Removed bitwise operators from the language
TI doesn't necessarily need bitwise operators but the real reason for removing them was because we couldn't overload the '|' operator, bitwise or otherwise, to also be the concat operator in lua due to it's nature to convert numbers to strings and therefore implicitly concatenate numbers as if they were strings. Changing that would have fundamentally have changed the nature of the language (apart from far from trivial to do).