Make ',' the only separator in a table constructor
A semicolon is allowed after every statement but silently ignored (treated as an empty statement, but therefore allowed as a separator character for convenience) but Lua allowed a semicolon also to be used as the separator in a table constructor where everywhere else only a comma can be used to separate variables, arguments, expressions, etc. Removed support of that making the comma separator the only separator in table constructors as well.