Accept '=' as equal operator in logical expression
In TI the equal sign is used as the binary equal operator in logical expressions. The language itself being that simple that it is not even actual overloading as one can't assign the result of a logical expression to a variable.
For backwards compatiblity we allow the equal sign as the equal operator in logical expressions used in the if, while and repeat conditions. Everywhere else, where before you couldn't use an equal operator in TI to begin with, you can now use the, proper, '==' operator.