Add support for TI specific IF and WHILE syntax
These changes add support for the 'classic' systax for IF and WHILE statements. When using these, function style, versions of IF and WHILE, the condition, the logical expression, needs to be enclosed in parenthesis, as if one were to pass it as an argument to a function, followed by a semicolon instead of the Lua style 'THEN' or 'DO' respectively. The whole statement is expected to use the same syntax style in that subsequent portions, the [ELSEIF*] [ELSE] and ENDIF are expected to use the same syntax, meaning the ElseIf conditions are also enclosed in parenthesis and all of these are followed by a semicolon as would the END at the end of a WHILE in that style.
Whilst using this 'classic' syntax the use of binary '=', '&', '%' and unary '~' operators, for 'equal', 'and', 'or' and 'not' respectively would be expected. A mix of classic and modern style is explicitly not permitted.