Add support for TI specific IF and WHILE syntax
These changes accept the 'THEN' to be ommitted if the IF condition is followed by a semicolon, subsequently expecting the rest of the IF [ELSEIF*] [ELSE] ENDIF to be using the same syntax and includes those semicolons. Same for WHILE where we'd except the 'DO' to be ommitted if the WHILE condition is followed by a simicolon, subsequently expecting the closing 'END' to be followed by a semicolon as well in that case.
Note that these changes don't replace the Lua style format where the THEN/DO would be expected but the semicolons are completely optional and treated as empty statements in statement blocks.