Replace band, bor, bxor, bnot, shl and shr operators
In TI we need the & (ampersand) for AND, % (percentage sign) for OR, ~ (tilde) for NOT and the | (pipe character) as the string concatenate operator.
Overloading in Lua wasn't an option due to it's nature to convert numbers to strings and therefore implicitly concatenate numbers as if they were strings. Changing that would have fundamentally have changed the nature of the language (apart from far from trivial to do).
Therefore we replaced the character based bitwise operators with their idiomatic names.